Separate models from protocols (#782)
