+++ BankWallet/BankWallet/Models/AdapterError.swift
@@ -1,0 +1,3 @@
+enum AdapterError: Error {
+    case unsupportedAccount
+}
