+++ BankWallet/BankWallet/Models/DefaultAccountType.swift
@@ -1,0 +1,4 @@
+enum DefaultAccountType {
+    case mnemonic(wordsCount: Int)
+    case eos
+}
