+++ BankWallet/BankWallet/Models/FeeRatePriority.swift
@@ -1,0 +1,7 @@
+enum FeeRatePriority: Int {
+    case lowest
+    case low
+    case medium
+    case high
+    case highest
+}
