+++ BankWallet/BankWallet/Models/SendTransactionError.swift
@@ -1,0 +1,4 @@
+enum SendTransactionError: Error {
+    case connection
+    case unknown
+}
