+++ BankWallet/BankWallet/Models/SyncMode.swift
@@ -1,0 +1,5 @@
+enum SyncMode: String {
+    case fast = "fast"
+    case slow = "slow"
+    case new = "new"
+}
