+++ BankWallet/BankWallet/UserInterface/LocalizedErrors.swift
@@ -39,3 +39,12 @@ extension EosKit.ValidationError: LocalizedError {
         }
     }
 }
+
+extension SendTransactionError: LocalizedError {
+    public var errorDescription: String? {
+        switch self {
+        case .connection: return "alert.no_internet".localized
+        case .unknown: return "alert.network_issue".localized
+        }
+    }
+}
